requirements.txt
beautifulsoup4==4.11.1
certifi==2022.9.24
charset-normalizer==2.1.1
numpy==1.23.4
pandas==1.5.1
python-dateutil==2.8.2
pytz==2022.5
requests==2.28.1
six==1.16.0
soupsieve==2.3.2.post1
urllib3==1.26.12
pymongo==4.3.2
newsapi-python== 0.2.6
pytest== 7.2.0
python-dotenv~=0.21.0